# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OllamaModelInfo": {
        "properties": {
          "modified_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "size_bytes"
        ],
        "title": "OllamaModelInfo",
        "type": "object"
      },
      "OllamaSettingsRead": {
        "properties": {
          "expose_to_sessions": {
            "title": "Expose To Sessions",
            "type": "boolean"
          }
        },
        "required": [
          "expose_to_sessions"
        ],
        "title": "OllamaSettingsRead",
        "type": "object"
      },
      "OllamaSettingsWrite": {
        "properties": {
          "expose_to_sessions": {
            "title": "Expose To Sessions",
            "type": "boolean"
          }
        },
        "required": [
          "expose_to_sessions"
        ],
        "title": "OllamaSettingsWrite",
        "type": "object"
      },
      "OllamaStatus": {
        "properties": {
          "available_models": {
            "items": {
              "$ref": "#/components/schemas/OllamaModelInfo"
            },
            "title": "Available Models",
            "type": "array"
          },
          "base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Url"
          },
          "default_model": {
            "title": "Default Model",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "reachable": {
            "title": "Reachable",
            "type": "boolean"
          }
        },
        "required": [
          "enabled",
          "base_url",
          "reachable",
          "error",
          "default_model",
          "available_models"
        ],
        "title": "OllamaStatus",
        "type": "object"
      },
      "OllamaTransformRequest": {
        "properties": {
          "mode": {
            "enum": [
              "mask",
              "anonymize",
              "summarize"
            ],
            "title": "Mode",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "mode"
        ],
        "title": "OllamaTransformRequest",
        "type": "object"
      },
      "OllamaTransformResponse": {
        "properties": {
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "model_used": {
            "title": "Model Used",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "mode",
          "model_used"
        ],
        "title": "OllamaTransformResponse",
        "type": "object"
      },
...
      "ProjectRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_branch": {
            "title": "Default Branch",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "provider": {
            "enum": [
              "github",
              "forgejo"
            ],
            "title": "Provider",
            "type": "string"
          },
          "repo_full_name": {
            "title": "Repo Full Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workspace_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Path"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "id",
          "name",
          "slug",
          "provider",
          "repo_full_name",
          "default_branch",
          "workspace_path"
        ],
        "title": "ProjectRead",
        "type": "object"
      },
...
      "RemoteRepoCreate": {
        "description": "Payload pour créer un nouveau repo distant (et le projet associé).",
        "properties": {
          "create_project": {
            "default": true,
            "title": "Create Project",
            "type": "boolean"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "private": {
            "default": true,
            "title": "Private",
            "type": "boolean"
          },
          "provider": {
            "enum": [
              "github",
              "forgejo"
            ],
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "name"
        ],
        "title": "RemoteRepoCreate",
        "type": "object"
      },
      "RemoteRepoRead": {
        "description": "Repo distant retourné par GitProvider.list_repos.",
        "properties": {
          "clone_url": {
            "title": "Clone Url",
            "type": "string"
          },
          "default_branch": {
            "title": "Default Branch",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "is_private": {
            "title": "Is Private",
            "type": "boolean"
          }
        },
        "required": [
          "full_name",
          "default_branch",
          "clone_url",
          "description",
          "is_private"
        ],
        "title": "RemoteRepoRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Eldir",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ollama/settings": {
      "get": {
        "operationId": "get_ollama_settings_api_v1_ollama_settings_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OllamaSettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ollama Settings",
        "tags": [
          "ollama"
        ]
      },
      "put": {
        "operationId": "set_ollama_settings_api_v1_ollama_settings_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OllamaSettingsWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OllamaSettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Ollama Settings",
        "tags": [
          "ollama"
        ]
      }
    },
    "/api/v1/ollama/status": {
      "get": {
        "operationId": "get_status_api_v1_ollama_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OllamaStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Status",
        "tags": [
          "ollama"
        ]
      }
    },
    "/api/v1/ollama/transform": {
      "post": {
        "operationId": "transform_api_v1_ollama_transform_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OllamaTransformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OllamaTransformResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transform",
        "tags": [
          "ollama"
        ]
      }
    },
...
    "/api/v1/providers/{provider}/repos": {
      "get": {
        "operationId": "list_remote_repos_api_v1_providers__provider__repos_get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RemoteRepoRead"
                  },
                  "title": "Response List Remote Repos Api V1 Providers  Provider  Repos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Remote Repos",
        "tags": [
          "providers"
        ]
      },
      "post": {
        "description": "Crée un nouveau repo côté provider et, optionnellement, le clone en projet.",
        "operationId": "create_remote_repo_api_v1_providers__provider__repos_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteRepoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ProjectRead"
                    },
                    {
                      "$ref": "#/components/schemas/RemoteRepoRead"
                    }
                  ],
                  "title": "Response Create Remote Repo Api V1 Providers  Provider  Repos Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Remote Repo",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}